Ajuste na estrutura e implementação dos botões de visualizar o cadastro, limpar e voltar
