+++ src/components/WeeksCalculator.tsx
@@ -1,19 +1,28 @@
-import React, { useState } from 'react';
+import React from 'react';
 
 const SectionTitle: React.FC<{ title: string }> = ({ title }) => (
   <h2 className="text-2xl font-bold text-center text-[#1a5276] mb-6">{title}</h2>
 );
 
-function WeeksCalculator() {
 
-  const [lastMenstruation, setLastMenstruation] = useState('');
-  
-  const [gestationalWeeks, setGestationalWeeks] = useState<number | null>(null);
+interface WeeksCalculatorProps {
+  lastMenstruation: string;
+  setLastMenstruation: (value: string) => void;
+  gestationalWeeks: number | null;
+  setGestationalWeeks: (value: number | null) => void;
+  message: string;
+  setMessage: (value: string) => void;
+}
 
-  const [message, setMessage] = useState('');
-
-  const handleCalculate = () => {
-
+function WeeksCalculator({
+  lastMenstruation,
+  setLastMenstruation,
+  gestationalWeeks,
+  setGestationalWeeks,
+  message,
+  setMessage,
+}: WeeksCalculatorProps) {
+    const handleCalculate = () => {
     if (!lastMenstruation) {
       setMessage('Por favor, insira a data da última menstruação.');
       setGestationalWeeks(null); 
@@ -49,11 +58,11 @@ function WeeksCalculator() {
               type="date"
               id="last-menstruation"
               value={lastMenstruation}
-              onChange={(e) => setLastMenstruation(e.target.value)}
+              onChange={(e) => setLastMenstruation(e.target.value)} 
               className="p-2 border border-gray-300 rounded"
             />
           </div>
-          <button
+          <button 
             type="button" 
             onClick={handleCalculate} 
             className="w-full sm:w-auto mt-4 sm:mt-0 self-end bg-[#1a5276] text-white py-2 px-4 rounded hover:bg-[#0e3040] transition-colors"
